+++ src/js/components/lists/articles/state/actions.js
@@ -4,7 +4,7 @@ import get from 'lodash.get';
 
 // get index from heroku, get jsons from s3 directly
 const baseS3Url = 'https://log-notes-assets.s3.amazonaws.com/';
-const baseHerokuUrl = 'https://log-notes-assets-api.herokuapp.com/';
+const baseHerokuUrl = 'https://log-notes-assets-api.herokuapp.com/json/';
 const config = { header: { 'Content-Type': 'application/json' } };
 
 const fetchArticle = articleId => {
@@ -15,7 +15,7 @@ const fetchArticle = articleId => {
         return dispatch({ article: payload.data, type: _types.SUCCESSFUL_SINGLE_ARTICLE_GET_REQUEST });
       })
       .catch(error => {
-        console.log('could not process data', error);
+        console.log('article fetch error:', error);
         return dispatch({ error: error, type: _types.FAILED_SINGLE_ARTICLE_GET_REQUEST });
       })
       .finally(() => {
@@ -41,16 +41,16 @@ const startListGetRequest = () => {
 
 export const fetchArticles = () => {
   return dispatch => {
-    const url = baseHerokuUrl + 'index';
+    const url = baseHerokuUrl + 'object/index';
     dispatch(startListGetRequest());
     return axios.get(url, config)
       .then(payload => {
-        const list = get(payload, 'data.payload.list', []);
+        const list = get(payload, 'data.payload', []);
         processArticleListPayload(dispatch, list);
         return dispatch({ index: list, type: _types.SUCCESSFUL_ARTICLE_LIST_GET_REQUEST });
       })
       .catch(error => {
-        console.log('error::', error);
+        console.log('index fetch error:', error);
         return dispatch({ error: error, type: _types.FAILED_ARTICLE_LIST_GET_REQUEST });
       })
       .finally(() => {