+++ src/js/libs/date.js
@@ -1,9 +1,29 @@
 const generateDateString = () => {
-  const date = new Date();
-  const year = date.getFullYear();
-  const month = (date.getMonth() < 9) ? '0' + (date.getMonth() + 1) : (date.getMonth() + 1);
-  const day = (date.getDate() < 10) ? '0' + date.getDate() : date.getDate();
-  return year + '-' + month + '-' + day;
+  const currDate = new Date();
+  return getYear(currDate) + '-' + getMonth(currDate) + '-' + getDay(currDate);
 };
 
-export { generateDateString };
+const getFullTimeStampString = () => {
+  const currDate = new Date();
+  const yyyyMMdd = getYear(currDate) + getMonth(currDate) + getDay(currDate);
+  const hhmmss = padNumber(currDate.getHours()) + padNumber(currDate.getMinutes()) + padNumber(currDate.getSeconds());
+  return yyyyMMdd + hhmmss + currDate.getMilliseconds();
+};
+
+const getYear = (date) => {
+  return date.getFullYear();
+};
+
+const getMonth = (date) => {
+  return padNumber(date.getMonth() + 1);
+};
+
+const getDay = (date) => {
+  return padNumber(date.getDate());
+};
+
+const padNumber = (value) => {
+  return ('' + value).padStart(2, '0');
+};
+
+export { generateDateString, getFullTimeStampString };